Observer Pattern. - 변경에 관심있는 대상(Subject)와 Subject에 변경이 생기면 notify 해달라는 걸로 역할, 책임 분리함. - Java9에서는 java.util.Observer가 deprecated 가 되었음.
This class and the Observer interface have been deprecated. The event model supported by Observer and Observable is quite limited, the order of notifications delivered by Observable is unspecified, and state changes are not in one-for-one correspondence with notifications. For a richer event model, consider using the java.beans package. For reliable and ordered messaging among threads, consider using one of the concurrent data structures in the java.util.concurrent package. For reactive streams style programming, see the java.util.concurrent.Flow API.